+++ DK/DKFoundation/DKStream.h
@@ -2,7 +2,7 @@
 //  File: DKStream.h
 //  Author: Hongtae Kim (tiff2766@gmail.com)
 //
-//  Copyright (c) 2004-2014 Hongtae Kim. All rights reserved.
+//  Copyright (c) 2004-2015 Hongtae Kim. All rights reserved.
 //
 
 #pragma once
@@ -20,7 +20,7 @@ namespace DKFoundation
 	class DKStream
 	{
 	public:
-		typedef long long Position;
+		typedef int64_t Position;
 
 		DKStream(void) {}
 		virtual ~DKStream(void) {}
